refactor: consolidate api services
