+++ apps/web/src/App.tsx
@@ -6,7 +6,7 @@ import Draft from './routes/draft'
 import { useState } from 'react'
 import EventDetail from './components/EventDetail'
 import InviteCollaboratorsModal from './components/InviteCollaboratorsModal'
-import { createDraftItinerary } from './lib/api'
+import { createDraftItinerary } from './lib/services/itinerary'
 import Calendar from './components/Calendar'
 import EventList from './components/EventList'
 import MapView from './components/MapView'
