--- apps/web/src/api/itinerary.ts
@@ -1,6 +1,0 @@
-export async function createDraftItinerary(day: number) {
-  // Stubbed API call
-  console.log('createDraftItinerary for day', day)
-  return { success: true }
-}
-
